pyproject.toml
[project]
name = "assetops-kg"
version = "0.1.0"
description = "Industrial Asset Operations Knowledge Graph — extending IBM AssetOpsBench with graph+vector+optimization"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "samyama>=0.6.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "sentence-transformers>=2.2.0",
    "fastmcp>=0.1.0",
    "click>=8.0",
    "rich>=13.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
